chore: move rp.writeTrampoline into proper file (#18)
